refactor: move options to policies module
